requirements.txt
.
anyio==3.6.2
async-timeout==4.0.2
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
-e git+https://github.com/nacifyas/cm-project@70598b36ee3028433ecff90576d2f4bd78f157ea#egg=cm_project
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.2
h11==0.14.0
hiredis==2.2.3
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
more-itertools==8.14.0
orjson==3.8.12
passlib==1.7.4
pptree==3.1
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-ulid==1.1.0
PyYAML==6.0
redis==4.5.5
redis-om==0.1.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
types-pyOpenSSL==23.1.0.3
types-redis==4.5.5.2
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
